refactored /services to match /routes, fixed init_db.py
