+++ backend/api/__init__.py
@@ -4,7 +4,6 @@ from flask_sqlalchemy import SQLAlchemy
 from flask_cors import CORS
 from flask_migrate import Migrate  # Import Migrate
 from dotenv import load_dotenv
-import os
 
 # Initialize extensions
 ma = Marshmallow()
