+++ backend/init_db.py
@@ -1,5 +1,8 @@
-from api import db, app
+from api import db, create_app
 from api.models import *
+
+app = create_app()
+
 # Create Postgres tables
 def create_tables():
     with app.app_context():
