+++ frontend/src/pages/instructor/assignments/index.js
@@ -2,7 +2,7 @@ import { CopyFilled, DownloadOutlined } from "@ant-design/icons";
 import { Button, Form, PageHeader, Popover, Space, Typography } from "antd";
 import { useCallback, useState } from "react";
 import { useParams } from "react-router-dom";
-import { createAssignment } from "../../../services/course";
+import { createAssignment } from "../../../services/assignment";
 import Assignments from "./Assignments";
 import CreateAssignment from "./CreateAssignment";
 import DuplicateAssignmentModal from "./DuplicateAssignmentModal";
