+++ frontend/src/services/assignment.js
@@ -1,5 +1,13 @@
 import service from ".";
 
+export async function createAssignment(params) {
+  return service("create_assignment", params, "post");
+}
+
+export async function createExtension(params) {
+  return service("create_extension", params, "post");
+}
+
 export async function getAssignment(params) {
   return service("get_assignment", params);
 }
@@ -10,8 +18,4 @@ export async function getExtension(params) {
 
 export async function updateAssignment(params) {
   return service("update_assignment", params, "post");
-}
-
-export async function createExtension(params) {
-  return service("create_extension", params, "post");
 }
