--- frontend/src/services/configureAutograder.js
@@ -1,5 +1,0 @@
-import service from ".";
-
-export async function uploadAssignmentAutograder(params) {
-  return service("upload_assignment_autograder", params, "post");
-}
